RCON Forum Bug
1. Added in a message that is sent to the logs and RCON console to show
that the UpdatePlayerDetails method has been called.

Currently there is an issue in the loading of the profile and tribe
files that prevents some server admins using the RCON window seeing any
user that is NOT online. It appears as though there is an issue with
ArkData timing out when trying to load a profile or tribe profile file
and the UpdatePlayerDetails is constantly being called.

Users are not aware that anything is wrong until after a while their
user list only shows the ONLINE players. This is NOT a fix, just an
information message to show that this issue is occurring.

I will download the ArkData code, and try to debug the issue and apply a
fix.