RCON Sort Option
Added a new sort option - By Last Updated. This option sorts the players
list on the last time the profile file was updated, in descending order.

Should be useful for server admins to determine which profiles can be
marked for removal as the older profiles will be listed at the bottom.

NOTE: This field is only set if access to the profile files was
successful.