Forum User Request
1. I have removed the code that sets the ServerIP if the Let ARK choose
option was previously selected.

NOTE: When the Let ARK choose option is selected the Mutihome settings
in the GameUserSettings.ini file are cleared. The server can be started
when these settings are cleared and also connected to.

I am unsure why this assignment was included, but when ARK is reopened,
the recommended IPAddress is assigned and the ini files are updated.

This change is to prevent the ini file from being updated when ASM
starts when the Let Ark choose is selected.