Auto-Update & Mod Update (Interim Checkin)
***** These changes were getting to big to keep checked out *****

* [NEW] Auto Update and Auto-Restart have been completely re-written.
* [NEW] Manual server update now includes mod updating.
* [NEW] Auto server update now includes mod updating, this includes the
map mod.
* [NEW] Auto Update and Auto Restart now check for online players and
will shutdown the server after the grace period or when there are no
online players.
* [CHANGE] Settings Window - New settings added for Auto Update and
shutdown functionality.
* [CHANGE] Automatic Management Section - changed for the new auto
update and auto restart functionality.
* [NEW] Added a new Mod Details button (?) next to the mod id list to
display a list of the mods. Any outdated mods are shown in orange.
* [NEW] Added a new Mod Update button (arrow) next to the mod id list to
update the mods only. This does not update the server.
* [NEW] Added a new reset button near the Map name to reset the value
back to TheCenter map.
* [BUG] Fixed a bug where FIPS enabled machine would crash when creating
scheduled tasks.
* [NEW] Administration section -> Server options - Added missing command
line options.
* [NEW] Administration section - added the WebAlerts section.
* [NEW] Administration section -> mod options - added the new ARK 244.3
option for download/install/update mod when the server is started.
* [NEW] Stop Button - when your server is running and you click the stop
button, a new dialog window is displayed allowing you to STOP your
server, or SHUTDOWN your server with a countdown.
* [NEW] New log files are created when the auto update and auto restart
processes are run.