Manual Server Installation Changes
1. When performing a new server install, will now check if auto-update
is enabled and cache folder set. If yes to both, then will copy the
files from the local cache first.
2. After the copy from local cache, a server validation is performed to
ensure it is up-to-date and valid.
3. When coping the server files and mod files, I now show a '.'
character that represents a folder being copied. This way it shows
progress. Each line of '.' will be 200 long, so each line represents 200
folders copied.
4. I have also changed the mod copy call to use a thread, so that the
progress window is updated properly.